renamed text_display to view, and moved srs into it
